Initial cut at the standard box
It will serve as a base for language/technology-specific VMs. So we will take standard box as a base
and then in addition to that provision more stuff by overriding Chef node attributes list, for example,
to provide N rubies or node.js versions instead of just one.

Per discussion with Josh K