Use Ruby 1.9.2 for Chef (read on)
Chef 0.10.8, that is the most recent version at the moment, depends on JSON gem
1.6.1. That gem seems to be built using Ruby 1.9 so Ruby 1.8 fails to install it
(thanks, buggy Ruby 1.9 YAML engine and RubyGems!). This is a known issue in the
Ruby community and as far as I know, even RubyGems 1.8.12 does not solve it. In
addition, it is probably impossible to solve for gems that were already released. We have had this issue with amqp gem, for example.

All this means that we need to run Chef on 1.9. Veewee installs a completely
separate Ruby into /opt *just* to be used by Vagrant. It  will never interfere
with our RVM Rubies, system Ruby or anything else. So it sounds like a good idea.

In addition, I removed puppet installation from the definition. Puppet is a nice
tool but we use Chef and are deeply rooted in it.